% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtconfirm.R
\name{run_rtconfirm}
\alias{run_rtconfirm}
\title{Run the model}
\usage{
run_rtconfirm(country = "South_Korea", cut = 10)
}
\arguments{
\item{String}{defining country to run model on (default = South Korea)}
}
\value{
List of two plots to show model fit
}
\description{
Run the model
}
\examples{
plots <- run_rtconfirm()
}
